refactor set root package to io.jenkins.plugins.xygeni
